requirements.txt
pandas
streamlit
pyyaml
numpy